src/requirements.txt
blessings==1.6.1
certifi==2018.1.18
chardet==3.0.4
click==6.7
click-plugins==1.0.3
colorama==0.3.9
Flask==0.12.2
Flask-Limiter==1.0.1
idna==2.6
itsdangerous==0.24
Jinja2==2.10
limits==1.3
MarkupSafe==1.0
netaddr==0.7.19
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycrypto==2.6.1
python-ldap==3.0.0
requests==2.18.4
shodan==1.7.7
six==1.11.0
urllib3==1.22
Werkzeug==0.14.1
XlsxWriter==1.0.2
pysqlite==2.8.3
